#106 hotfix : 지역 목록 삭제시 대표지역 설정 추가 , 구조 리팩토링
