+++ secondhand/src/main/java/com/codesquad/secondhand/domain/member/controller/MemberController.java
@@ -10,11 +10,9 @@ import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -22,10 +20,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.codesquad.secondhand.domain.jwt.dto.request.ReissueTokenRequest;
 import com.codesquad.secondhand.domain.jwt.dto.response.ReissueTokenResponse;
 import com.codesquad.secondhand.domain.jwt.service.JwtService;
-import com.codesquad.secondhand.domain.member.dto.request.RegionRequest;
 import com.codesquad.secondhand.domain.member.dto.request.SignupRequest;
 import com.codesquad.secondhand.domain.member.dto.response.MemberInfoResponse;
-import com.codesquad.secondhand.domain.member.dto.response.MemberRegionResponse;
 import com.codesquad.secondhand.domain.member.dto.response.SignUpResponse;
 import com.codesquad.secondhand.domain.member.service.MemberService;
 
@@ -53,30 +49,6 @@ public class MemberController {
 		Long memberId = extractMemberId(request);
 		memberService.signOut(accessToken, memberId);
 		return ResponseEntity.ok(Collections.singletonMap("message", "로그아웃 성공"));
-	}
-
-	@PostMapping("/members/{memberId}/regions")
-	public ResponseEntity addRegion(@PathVariable Long memberId, @RequestBody RegionRequest regionRequest) {
-		memberService.addRegion(memberId, regionRequest);
-		return ResponseEntity.status(HttpStatus.CREATED).build();
-	}
-
-	@DeleteMapping("/members/{memberId}/regions")
-	public ResponseEntity deleteRegion(@PathVariable Long memberId, @RequestBody RegionRequest regionRequest) {
-		memberService.deleteRegion(memberId, regionRequest);
-		return ResponseEntity.noContent().build();
-	}
-
-	@PutMapping("/members/{memberId}/regions")
-	public ResponseEntity updateSelectedRegion(@PathVariable Long memberId, @RequestBody RegionRequest regionRequest) {
-		memberService.updateSelectedRegion(memberId, regionRequest);
-		return ResponseEntity.noContent().build();
-	}
-
-	@GetMapping("/members/{memberId}/regions")
-	public ResponseEntity<MemberRegionResponse> getRegions(@PathVariable Long memberId) {
-		MemberRegionResponse memberRegionResponse = memberService.getRegion(memberId);
-		return ResponseEntity.ok(memberRegionResponse);
 	}
 
 	//todo 이럴때 토큰에 있는 memberId랑 해당 memberId랑 같은지 수정해야함
